+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 + Generate token from token
++ Add Refreh Token route
+
 
 ## 0.11.0
 + Remove unused imports